refactor: add PasswordUtil for temporary password generation and update DoctorService to use it
