+++ src/common/utils/index.ts
@@ -1,3 +1,3 @@
 export * from './time.util';
 export * from './transaction.util';
-
+export * from './password.util';
