+++ src/common/utils/password.util.ts
@@ -1,0 +1,14 @@
+import * as crypto from 'crypto';
+
+export class PasswordUtil {
+  static generateTemporaryPassword(length: number = 16): string {
+    const charset = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*';
+    const randomBytes = crypto.randomBytes(length);
+    let password = '';
+    for (let i = 0; i < length; i++) {
+      password += charset[randomBytes[i] % charset.length];
+    }
+    return password;
+  }
+}
+
