+++ src/doctor/doctor.service.ts
@@ -1,10 +1,9 @@
 import { Injectable, ConflictException, NotFoundException, Logger } from '@nestjs/common';
 import { PrismaService } from '../prisma';
-import { CreateDoctorDto, QueryDoctorDto, DoctorResponseDto } from './dto';
+import { CreateDoctorDto, QueryDoctorDto } from './dto';
 import { UserService } from '../user';
-import { TransactionUtil } from '../common/utils';
+import { TransactionUtil, PasswordUtil } from '../common/utils';
 import * as bcrypt from 'bcrypt';
-import * as crypto from 'crypto';
 import { UserRole } from 'src/common/constants';
 import { DoctorStatus, Prisma } from '@prisma/client';
 import { PaginatedResponse } from '../common/dto';
@@ -20,7 +19,7 @@ export class DoctorService {
 
   async create(createDoctorDto: CreateDoctorDto) {
     await this.validateData(createDoctorDto);
-    const temporaryPassword = this.generateTemporaryPassword();
+    const temporaryPassword = PasswordUtil.generateTemporaryPassword();
     const hashedPassword = await bcrypt.hash(temporaryPassword, 10);
     this.logger.log(`Generated temporary password for doctor: ${createDoctorDto.email}`);
     const doctor = await TransactionUtil.executeInTransaction(this.prisma, async (tx) => {
@@ -100,19 +99,25 @@ export class DoctorService {
       temporaryPassword,
     };
   }
-
-  private generateTemporaryPassword(): string {
-    const length = 16;
-    const charset = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%^&*';
-    const randomBytes = crypto.randomBytes(length);
-    let password = '';
-    for (let i = 0; i < length; i++) {
-      password += charset[randomBytes[i] % charset.length];
+  
+  async findAll(query: QueryDoctorDto) {
+    const where = this.buildFilterQuery(query);
+    const [doctors, total] = await Promise.all([
+      this.prisma.doctor.findMany({
+        where,
+        ...query.getPrismaParams(),
+        orderBy: query.getPrismaSortParams(),
+        include: this.getDoctorIncludes(),
+      }),
+      this.prisma.doctor.count({ where }),
+    ]);
+    return {
+      data: doctors,
+      total,
     }
-    return password;
   }
 
-  async findAll(query: QueryDoctorDto) {
+  private buildFilterQuery(query: QueryDoctorDto): Prisma.DoctorWhereInput {
     const where: Prisma.DoctorWhereInput = {};
     if (query.name) {
       where.user = {
@@ -128,69 +133,32 @@ export class DoctorService {
     if (query.status) {
       where.status = query.status;
     }
-    const [doctors, total] = await Promise.all([
-      this.prisma.doctor.findMany({
-        where,
-        ...query.getPrismaParams(),
-        orderBy: query.getPrismaSortParams(),
-        include: {
-          user: {
-            select: {
-              id: true,
-              email: true,
-              username: true,
-              phone: true,
-              fullName: true,
-              avatar: true,
-              address: true,
-              role: true,
-            },
-          },
-          primarySpecialty: {
-            select: {
-              id: true,
-              name: true,
-              description: true,
-              isActive: true,
-            },
-          },
-        },
-      }),
-      this.prisma.doctor.count({ where }),
-    ]);
-    const doctorResponses: DoctorResponseDto[] = doctors.map((doctor) => ({
-      id: doctor.id,
-      userId: doctor.userId,
-      user: {
-        id: doctor.user.id,
-        email: doctor.user.email,
-        username: doctor.user.username ?? undefined,
-        phone: doctor.user.phone ?? undefined,
-        fullName: doctor.user.fullName ?? undefined,
-        avatar: doctor.user.avatar ?? undefined,
-        address: doctor.user.address ?? undefined,
-        role: doctor.user.role,
-      },
-      primarySpecialtyId: doctor.primarySpecialtyId,
-      primarySpecialty: {
-        id: doctor.primarySpecialty.id,
-        name: doctor.primarySpecialty.name,
-        description: doctor.primarySpecialty.description ?? undefined,
-        isActive: doctor.primarySpecialty.isActive,
-      },
-      subSpecialty: doctor.subSpecialty ?? undefined,
-      professionalTitle: doctor.professionalTitle ?? undefined,
-      yearsOfExperience: doctor.yearsOfExperience,
-      consultationFee: doctor.consultationFee,
-      bio: doctor.bio ?? undefined,
-      status: doctor.status,
-      createdAt: doctor.createdAt,
-      updatedAt: doctor.updatedAt,
-    }));
+    return where;
+  }
+
+  private getDoctorIncludes() {
     return {
-      data: doctorResponses,
-      total,
-    }
+      user: {
+        select: {
+          id: true,
+          email: true,
+          username: true,
+          phone: true,
+          fullName: true,
+          avatar: true,
+          address: true,
+          role: true,
+        },
+      },
+      primarySpecialty: {
+        select: {
+          id: true,
+          name: true,
+          description: true,
+          isActive: true,
+        },
+      },
+    };
   }
   private async validateData(dto: CreateDoctorDto): Promise<void> {
     const [existingUser, specialty] = await Promise.all([
